+++ internal/entity/rivalScoreDataLog.go
@@ -31,7 +31,7 @@ type RivalScoreDataLog struct {
 	PlayCount  int32
 	ClearCount int32
 	Option     int32
-	Seed       int32
+	Seed       int64
 	Random     int32
 	State      int32
 }
